[IMP] purchase_delivery_split_date: Split one model per file
