+++ purchase_delivery_split_date/models/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import purchase
+from . import purchase_order, purchase_order_line, stock_picking
