+++ purchase_delivery_split_date/models/purchase_order_line.py
@@ -8,6 +8,7 @@ from odoo import api, fields, models
 
 
 class PurchaseOrderLine(models.Model):
+
     _inherit = "purchase.order.line"
 
     @api.model
@@ -103,60 +104,3 @@ class PurchaseOrderLine(models.Model):
             ):
                 line.date_planned = date_planned_by_record[line.id]
         return res
-
-
-class PurchaseOrder(models.Model):
-    _inherit = "purchase.order"
-
-    def _check_split_pickings(self):
-        for order in self:
-            moves = self.env["stock.move"].search(
-                [
-                    ("purchase_line_id", "in", order.order_line.ids),
-                    ("state", "not in", ("cancel", "done")),
-                ]
-            )
-            pickings = moves.mapped("picking_id")
-            pickings_by_date = {}
-            for pick in pickings:
-                pickings_by_date[pick.scheduled_date.date()] = pick
-
-            order_lines = moves.mapped("purchase_line_id")
-            date_groups = groupby(
-                order_lines, lambda l: l._get_group_keys(l.order_id, l)
-            )
-            for key, lines in date_groups:
-                date_key = fields.Date.from_string(key[0]["date_planned"])
-                for line in lines:
-                    for move in line.move_ids:
-                        if move.state in ("cancel", "done"):
-                            continue
-                        if (
-                            move.picking_id.scheduled_date.date() != date_key
-                            or pickings_by_date[date_key] != move.picking_id
-                        ):
-                            if date_key not in pickings_by_date:
-                                copy_vals = line._first_picking_copy_vals(key, line)
-                                new_picking = move.picking_id.copy(copy_vals)
-                                pickings_by_date[date_key] = new_picking
-                            move._do_unreserve()
-                            move.picking_id = pickings_by_date[date_key]
-                            move.date_deadline = date_key
-                            move._action_assign()
-            for picking in pickings:
-                if len(picking.move_ids) == 0:
-                    picking.write({"state": "cancel"})
-
-
-class StockPicking(models.Model):
-    _inherit = "stock.picking"
-
-    def _update_picking_from_group_key(self, key):
-        """The picking is updated with data from the grouping key.
-        This method is designed for extensibility, so that other modules
-        can store more data based on new keys."""
-        for rec in self:
-            for key_element in key:
-                if "date_planned" in key_element.keys():
-                    rec.date = key_element["date_planned"]
-        return False
